pycerver-server/requirements.txt
pycerver
pymongo
pymongoose
pymongo[srv]
dnspython